API Type Declarations
Strengthening the self-type definitions for the internal parts of the library itself.

New things I learned today, `[].slice.call(an_iterator)` is another alternative to the spread operator. Also new to know that it's specific to ES2015, TS actually gave me an error as to not being able to use it when I set the target to ES5.

Also had to redefine a few variables from `var` to `let`, since they weren't being scoped to each `if` block they were defined in. It wasn't causing any errors before, since they would reassign them in each block anyways, but it's still a bit dangerous and forgetful had one not remembered to reassign it again for that loop.