Type Accuracy
Updating the user-defined types to match that of the return types and value types of the library's use of them. There were a few extra instances of things like `null`, or where a `number` should have been one of the `interface` types.

Was going to mention I've been wanting to work out this format for some time now, glad there was already a JS implementation of it. I went on a run today, and it got kind of messed up because my watch randomly crashed. Lost some of the metric data, so I wanted to see what it would look like to parse the GPX data of my Strava activity, since they allow you to access the file itself, super cool!