TS Migration
Moving over to TS, but I also still want to ensure that the test script can still be run before I do anything else. I also didn't want to change much of the files in the first commit since it makes it harder to diff when the file name also changed, usually Git doesn't do very well with that.